Move `report.jl`, add check for correctly formed reports
